Split cart state into focused modules
