+++ app/components/product/add-to-cart-button.tsx
@@ -13,7 +13,8 @@ import { useEffect, useRef, useState } from "react";
 import type { FetcherWithComponents } from "react-router";
 import { useMatches } from "react-router";
 import { Button } from "~/components/button";
-import { useCartFetcherSync, useCartStore } from "~/components/cart/store";
+import { useCartFetcherSync } from "~/components/cart/cart-sync";
+import { useCartStore } from "~/components/cart/store";
 import { Spinner } from "~/components/spinner";
 import { usePrefixPathWithLocale } from "~/hooks/use-prefix-path-with-locale";
 import { cn } from "~/utils/cn";
